requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-restful==0.5.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
psutil==5.9.6
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
redis==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.23.2
